Move generate_genesis_state to cumulus-primitives
