+++ test/service/src/genesis.rs
@@ -15,44 +15,15 @@
 // along with Cumulus.  If not, see <http://www.gnu.org/licenses/>.
 
 use codec::Encode;
-use cumulus_primitives::ParaId;
-use cumulus_test_runtime::Block;
+use cumulus_primitives::{ParaId, genesis::generate_genesis_state};
 use polkadot_primitives::v0::HeadData;
-use sp_runtime::traits::{Block as BlockT, Hash as HashT, Header as HeaderT, Zero};
+use sp_runtime::traits::{Block as BlockT};
+use cumulus_test_runtime::Block;
 
-pub fn generate_genesis_state(
-	chain_spec: &Box<dyn sc_service::ChainSpec>,
-) -> Result<Block, String> {
-	let storage = chain_spec.build_storage()?;
-
-	let child_roots = storage.children_default.iter().map(|(sk, child_content)| {
-		let state_root = <<<Block as BlockT>::Header as HeaderT>::Hashing as HashT>::trie_root(
-			child_content.data.clone().into_iter().collect(),
-		);
-		(sk.clone(), state_root.encode())
-	});
-	let state_root = <<<Block as BlockT>::Header as HeaderT>::Hashing as HashT>::trie_root(
-		storage.top.clone().into_iter().chain(child_roots).collect(),
-	);
-
-	let extrinsics_root =
-		<<<Block as BlockT>::Header as HeaderT>::Hashing as HashT>::trie_root(Vec::new());
-
-	Ok(Block::new(
-		<<Block as BlockT>::Header as HeaderT>::new(
-			Zero::zero(),
-			extrinsics_root,
-			state_root,
-			Default::default(),
-			Default::default(),
-		),
-		Default::default(),
-	))
-}
-
+/// Returns the initial head data for a parachain ID.
 pub fn initial_head_data(para_id: ParaId) -> HeadData {
 	let spec = Box::new(crate::chain_spec::get_chain_spec(para_id));
-	let block = generate_genesis_state(&(spec as Box<_>)).unwrap();
+	let block: Block = generate_genesis_state(&(spec as Box<_>)).unwrap();
 	let genesis_state = block.header().encode();
 	genesis_state.into()
 }
